requirements.txt
anthropic>=0.40.0
openai>=1.55.0
google-genai>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
